refactor(test-workflow.ts): move fetch-db-from-ci to firebase-utils and fix tsconfig
