+++ libs/firebase-utils/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/firebase-utils';
 export * from './lib/maintenance';
 export * from './lib/watermark';
+export * from './lib/test-workflow';
